Refactor example

+++ bluesky_widgets/qt/search_input.py
@@ -16,7 +16,7 @@ from ..components.search.search_input import LOCAL_TIMEZONE, secs_since_epoch
 def as_qdatetime(datetime):
     "Create QDateTime set as specified by datetime."
     return QDateTime.fromSecsSinceEpoch(
-        secs_since_epoch(datetime) - datetime.utcoffset() / timedelta(seconds=1)
+        int(secs_since_epoch(datetime) - datetime.utcoffset() / timedelta(seconds=1))
     )
